requirements.txt
open-clip-torch
easydict 
h5py
ninja
datasets
deepspeed
einops
huggingface_hub
matplotlib
# mmcv
# mmcv-full
# mmengine
numpy
open3d
# openmim
pandas
ruamel.yaml
scikit-learn
scipy
timm
tqdm
wandb